web5D/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modelcluster==5.0.2
django-taggit==1.3.0
django-toolbelt==0.0.1
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
l18n==2018.5
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
Wagtail-Multi-Image-Edit==0.0.5
webencodings==0.5.1
whitenoise==5.1.0
Willow==1.3
XlsxWriter==1.2.9
